pyproject.toml
[tool.poetry]
name = "knowde"
version = "0.0.0"                                  # poetry-dynamic-versioningでgit tagの値にに置換される
description = ""
authors = ["Shogo Goto <gotoadmn0605@gmail.com>"]
readme = "README.md"
repository = "https://github/com/shogogoto/knowde"

[tool.poetry.dependencies]
python = ">3.11.0,<3.14"
fastapi = { extras = ["standard"], version = "^0.115.6" }
neomodel = "5.4.4"
uvicorn = "^0.24.0.post1"
uuid6 = "^2023.5.2"
requests = "^2.31.0"
click = "^8.1.7"
tabulate = "^0.9.0"
makefun = "^1.15.2"
inflector = "^3.1.0"
networkx = "^3.2.1"
pydantic-partial = "^0.5.4"
compose-stubs = "^1.0.0"
compose = "^1.5.0"

more-itertools = "^10.3.0"
lark = "^1.2.2"
treelib = "^1.7.0"
levenshtein = "^0.26.1"
regex = "^2024.11.6"
numpy = "^2.2.2"
scipy = "^1.15.1"
taskipy = "^1.14.1"
japanera = "^2.1.2"
intervaltree = "^3.1.0"
edtf = "^5.0.0"
python-multipart = "^0.0.20"
pydantic-settings = "^2.7.1"
fastapi-users = { extras = ["oauth"], version = "^14.0.1" }
chardet = "^5.2.0"
ruff = "^0.11.2"
[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pre-commit = "^3.5.0"
requests-mock = "^1.11.0"
pytest-watch = "^4.2.0"
httpx = "^0.25.2"
pytest-unordered = "^0.5.2"

pytest-testmon = "^2.1.3"
pytest-env = "^1.1.5"
pyfakefs = "^5.7.4"
pytest-cov = "^6.0.0"


[tool.poetry.group.build.dependencies]
pyinstaller = "^6.11.1"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"

[tool.poetry-dynamic-versioning.substitution]
files = ["knowde/feature/cli/__init__.py"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.taskipy.tasks]
test = "pytest --cov"
ptw = "ptw -- --ff --testmon"                               # ispytest-watch
lint = "ruff check ."
lintfix = "ruff check . --fix"
pre-commit = "pre-commit install"
dev = "fastapi dev knowde/feature/api/__init__.py --reload"
# serve      = "uvicorn knowde:api --reload --host 0.0.0.0 --port 8000"
serve = "fastapi run knowde/feature/api/__init__.py"
build = "pyinstaller --noupx --optimize 1 --collect-all knowde --workpath ./standalone/build --specpath ./standalone --distpath ./standalone/dist  --onefile --name kn knowde/feature/cli/__init__.py"
compile = "nuitka --standalone --follow-imports --onefile --product-name=kn knowde/feature/cli.py"

[tool.poetry.scripts]
kn = "knowde:cli"
